Extract codegen out of clair-c2py into a shared directory
Move src/tools/clair-c2py/codegen/ to src/tools/codegen/ so the code
generation layer can be reused by future tools (e.g. clair-f2py) without
duplicating code. The codegen files are updated to reference wdata.hpp and
other clair-c2py headers via ../clair-c2py/ paths. CMakeLists picks up the
shared codegen sources via the "codegen/*.cpp" glob passed to add_tool.
