+++ src/tools/codegen/classes.hpp
@@ -1,4 +1,4 @@
 #pragma once
-#include "../wdata.hpp"
+#include "../clair-c2py/wdata.hpp"
 
 str_t codegen_cls(std::ostream &code, str_t const &cls_py_name, cls_info_t const &cls, str_t const &full_module_name);
