+++ src/tools/codegen/doc.cpp
@@ -1,5 +1,5 @@
 #include "doc.hpp"
-#include "../wdata.hpp"
+#include "../clair-c2py/wdata.hpp"
 #include <fmt/core.h>
 #include <fmt/format.h>
 #include <algorithm>
