+++ src/tools/codegen/fnt.cpp
@@ -1,7 +1,7 @@
 #include <algorithm>
 #include "fnt.hpp"
 #include "./utils.hpp"
-#include "../wdata.hpp"
+#include "../clair-c2py/wdata.hpp"
 #include <fmt/core.h>
 #include <fmt/format.h>
 using namespace fmt::literals;
