+++ src/tools/codegen/fnt_tools.cpp
