+++ src/tools/codegen/templates/module.txt
